[core] Move GeoJSON-specific parsing into geojson_source.cpp
